add vectorise as a function in python graph view
