+++ raphtory/src/python/graph/views/graph_view.rs
@@ -26,15 +26,20 @@ use crate::{
     prelude::*,
     python::{
         graph::{edge::PyEdges, index::GraphIndex, vertex::PyVertices},
+        packages::vectors::{spawn_async_task, DynamicVectorisedGraph, PyDocumentTemplate},
         types::repr::Repr,
         utils::{PyInterval, PyTime},
     },
+    vectors::vectorisable::Vectorisable,
     *,
 };
 use chrono::prelude::*;
 use itertools::Itertools;
-use pyo3::{prelude::*, types::PyBytes};
-use std::ops::Deref;
+use pyo3::{
+    prelude::*,
+    types::{PyBytes, PyFunction},
+};
+use std::{ops::Deref, path::PathBuf};
 
 impl IntoPy<PyObject> for MaterializedGraph {
     fn into_py(self, py: Python<'_>) -> PyObject {
@@ -297,6 +302,25 @@ impl PyGraphView {
     ///    GraphIndex - Returns a GraphIndex
     fn index(&self) -> GraphIndex {
         GraphIndex::new(self.graph.clone())
+    }
+
+    fn vectorise(
+        &self,
+        embedding: &PyFunction,
+        cache: Option<String>,
+        node_document: Option<String>,
+        edge_document: Option<String>,
+    ) -> DynamicVectorisedGraph {
+        let embedding: Py<PyFunction> = embedding.into();
+        let graph = self.graph.clone();
+        let cache = cache.map(|cache| PathBuf::from(cache));
+        let template = PyDocumentTemplate::new(node_document, edge_document);
+        spawn_async_task(move || async move {
+            let vectorised_graph = graph
+                .vectorise_with_template(Box::new(embedding.clone()), cache, template)
+                .await;
+            vectorised_graph
+        })
     }
 
     /// Get bincode encoded graph
